% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table2.R
\name{Table2}
\alias{Table2}
\title{Create journal style model output table.}
\usage{
Table2(model, data, catcols = NULL, esdigits = 2, output = NULL,
  pdigits = 2, eps = 0.001)
}
\arguments{
\item{model}{Object output from lm, glm or coxph}

\item{data}{A dataframe including the exposure variable.}

\item{esdigits}{Controlling the effect size digits. Default to 2.}

\item{output}{String of path to store the output word file. For example, 'Table2.doc' or 'Table2.rtf'}

\item{pdigits}{Controlling the significant p-value digits. Default to 2.}

\item{eps}{P-value tolerane. Those less than eps are formatted as "< [eps]". Default to 0.001}

\item{catcol}{A vector of factrozied categorical column names in character.  Default to NULL.}
}
\value{
If output is not specified, a dataframe will be returned. Otherwise, a rtf file will be saved in the specified path.
}
\description{
Create journal style model output table.
}
\details{
Table2 function

This function helps to create journal style "Table 2" (model output).
Currently the function only supports lm, glm and coxph objects.
\cr The output include effect size (95\% CI) along with p-values.
\cr The output will be saved in word file.
}
\examples{

data(mtcars)

log = glm(vs==1~mpg+am, family='binomial', data=mtcars)

  Variable     ES_CI              P_val
(Intercept)    0 (0, 0.03)        0.006
mpg            1.98 (1.2, 3.24)   0.007
am
  0            Ref                Ref
  1            0.05 (0, 1.14)     0.060
}
