% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{Table1}
\alias{Table1}
\title{Create journal style characteristic table}
\usage{
Table1(data, numcol = NULL, catcol = NULL, exp_var, output = NULL,
  overall = TRUE)
}
\arguments{
\item{data}{A dataframe including the exposure variable.}

\item{numcol}{A vector of numerical column names in character. Default to NULL.}

\item{catcol}{A vector of categorical column names in character. Default to NULL.}

\item{exp_var}{String of main exposure variable name}

\item{output}{String of path to store the output word file. For example, 'Table1.doc' or 'Table1.rtf'}

\item{overall}{Whether to add a column for overall subjects. Default to TRUE}
}
\value{
If output is not specified, a dataframe will be returned. Otherwise, a rtf file will be saved in the specified path.
}
\description{
Create journal style characteristic table
}
\details{
Table1 function

This function helps to create journal style "Table 1" for numerical/continuous and categorical columns.
\cr For numerical/continuous columns, the results will be mean(SD). P-value is computed from Kruskal-Wallis Rank Sum test statistic.
\cr For categorical columns, the results will be n(\%). P-value is computed from chi-square test statistic.
\cr The output will be saved in word file.
}
\examples{
df = data.frame(
a = sample(1:100, 100, TRUE),
b = sample(c('Y', 'N', 'UNK'), 100, TRUE, prob=c(0.5, 0.3, 0.2)),
c = sample(1:100, 100, TRUE),
d = sample(c('Exposed', 'Unexposed'), 100, TRUE, prob=c(0.6, 0.4)))

Table1(df, c('a', 'c'), 'b', 'd', overall=TRUE)
name Overall (n=100) Exposed (n=60) Unexposed (n=40)     P_val
b: N         28 (28)        18 (30)          10 (25)     0.804
b: UNK       27 (27)        15 (25)          12 (30)
b: Y         45 (45)        27 (45)          18 (45)
a            54.2 (28.6)    52.2 (28.3)      57.2 (29.2) 0.408
c            50.1 (31)      50.1 (29.7)      50.1 (33.3) 0.935
}
